requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
pydantic==1.10.17
python-dotenv==1.0.1
